Add compiled universalJavaApplicationStub as option
This is my proposal to have the universalJavaApplicationStub compiled to a binayr before codesigning so projects that depend on JFileCHooser will have a properly working JFileChooser that can access the user's Documents/Downloads/Desktop folders.

I would suggest to offer this as an option in javapackager, since it needs an extra dependency. You need to have the shell compiler shc installed and available on the commandline using:

https://github.com/neurobin/shc

Download it with
git clone https://github.com/neurobin/shc.git
and..
./configure
make
sudo make install

This is working properly on both intel and arm macs for my project and is completely inline with the information from universalJavaApplicationStub itself, so I can only assume it will work properly for all other projects with the current macOS security and privacy settings on the newest macOS versions.